backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pillow==12.0.0
python-multipart==0.0.6